Move endpoints handlers to separate package
