+++ cogboard-app/src/main/kotlin/com/cognifide/cogboard/CogboardConstants.kt
@@ -31,7 +31,7 @@ class CogboardConstants {
         const val EVENT_SEND_MESSAGE_TO_WEBSOCKET = "cogboard.websocket.message"
         const val EVENT_UPDATE_WIDGET_CONFIG = "cogboard.config.widget.update"
         const val EVENT_DELETE_WIDGET_CONFIG = "cogboard.config.widget.delete"
-        const val EVENT_UPDATE_ENDPOINTS_CONFIG = "cogboard.config.endpoints.add"
+        const val EVENT_UPDATE_ENDPOINTS_CONFIG = "cogboard.config.endpoints.update"
         const val EVENT_CONFIG_SAVE = "cogboard.config.save"
         const val EVENT_HTTP_GET = "cogboard.httpclient.get"
         const val EVENT_HTTP_CHECK = "cogboard.httpclient.check"
