+++ knotx/conf/openapi.yaml
@@ -47,12 +47,12 @@ paths:
       responses:
         default:
           description: Endpoint list
-  /api/endpoints/add:
+  /api/endpoints/update:
     post:
       operationId: endpoints-update-handler
       responses:
         default:
-          description: Endpoints Config Update Handler
+          description: Endpoint Update Handler
   /api/config/save:
     post:
       operationId: config-save-handler
